Address review: withdraw -Wconversion and the accompanying casts
Per review feedback, all changes that existed only to silence
-Wconversion are withdrawn:

- CMakeLists.txt: back to -Wall -Wextra for GCC; the libyuv
  -Wno-conversion scoping goes away with the flag.
- apps/shared/iccmaker.c, src/utils.c, src/reformat.c,
  src/reformat_libyuv.c, src/stream.c: revert to the original
  expressions; masked byte extraction provably fits the target type,
  so the explicit casts were unnecessary.
- src/gainmap.c, src/read.c: revert the redundant float casts; when
  the first operand of an arithmetic expression is float, the other
  operand is implicitly converted, so one explicit cast (or none,
  when the first operand is already float) suffices.

The actual truncation fixes and their regression test are unchanged:
sato input item count uint8_t -> uint32_t, 64-bit OBU size
computation with bound check, and a1lx layer size bound check.

Verified: 0 warnings with the upstream GCC flags, ctest 51/51.

+++ apps/shared/iccmaker.c
@@ -148,24 +148,24 @@ static uint32_t readLittleEndianU32(const uint8_t * data)
 
 static void writeLittleEndianU32(uint8_t * data, uint32_t value)
 {
-    data[0] = (uint8_t)((value >> 0) & 0xff);
-    data[1] = (uint8_t)((value >> 8) & 0xff);
-    data[2] = (uint8_t)((value >> 16) & 0xff);
-    data[3] = (uint8_t)((value >> 24) & 0xff);
+    data[0] = (value >> 0) & 0xff;
+    data[1] = (value >> 8) & 0xff;
+    data[2] = (value >> 16) & 0xff;
+    data[3] = (value >> 24) & 0xff;
 }
 
 static void writeBigEndianU16(uint8_t * data, uint16_t value)
 {
-    data[0] = (uint8_t)((value >> 8) & 0xff);
-    data[1] = (uint8_t)((value >> 0) & 0xff);
+    data[0] = (value >> 8) & 0xff;
+    data[1] = (value >> 0) & 0xff;
 }
 
 static void writeBigEndianU32(uint8_t * data, uint32_t value)
 {
-    data[0] = (uint8_t)((value >> 24) & 0xff);
-    data[1] = (uint8_t)((value >> 16) & 0xff);
-    data[2] = (uint8_t)((value >> 8) & 0xff);
-    data[3] = (uint8_t)((value >> 0) & 0xff);
+    data[0] = (value >> 24) & 0xff;
+    data[1] = (value >> 16) & 0xff;
+    data[2] = (value >> 8) & 0xff;
+    data[3] = (value >> 0) & 0xff;
 }
 
 static avifBool putS15Fixed16(uint8_t * data, double value)
